Add files via upload
I left only what is needed from the original code and commented out everything else. Also, I added default args which are used in the implementation of the paper.

TODO: Check whether fp is correct, simplify the code, convert some lines that are only compatible with the older version of tf to tf2.0 compatible codes.